follow-up to telemetry collection PR
